# Conflicts:
#	eperusteet-amosaa-service/src/main/java/fi/vm/sade/eperusteet/amosaa/config/MetricsConfig.java
#	eperusteet-amosaa-service/src/main/java/fi/vm/sade/eperusteet/amosaa/config/MvcConfiguration.java
#	eperusteet-amosaa-service/src/main/java/fi/vm/sade/eperusteet/amosaa/config/SwaggerConfig.java
#	eperusteet-amosaa-service/src/main/java/fi/vm/sade/eperusteet/amosaa/domain/liite/version/JpaWithVersioningRepository.java
#	eperusteet-amosaa-service/src/main/java/fi/vm/sade/eperusteet/amosaa/domain/liite/version/JpaWithVersioningRepositoryFactoryBean.java
#	eperusteet-amosaa-service/src/main/java/fi/vm/sade/eperusteet/amosaa/domain/liite/version/JpaWithVersioningRepositoryImpl.java
#	eperusteet-amosaa-service/src/main/java/fi/vm/sade/eperusteet/amosaa/repository/kayttaja/KayttajaRepository.java
#	eperusteet-amosaa-service/src/main/java/fi/vm/sade/eperusteet/amosaa/repository/koulutustoimija/KoulutustoimijaRepository.java
#	eperusteet-amosaa-service/src/main/java/fi/vm/sade/eperusteet/amosaa/repository/koulutustoimija/TermistoRepository.java
#	eperusteet-amosaa-service/src/main/java/fi/vm/sade/eperusteet/amosaa/repository/koulutustoimija/TiedoteRepository.java
#	eperusteet-amosaa-service/src/main/java/fi/vm/sade/eperusteet/amosaa/repository/liite/LiiteRepository.java
#	eperusteet-amosaa-service/src/main/java/fi/vm/sade/eperusteet/amosaa/repository/tutkinnonosa/TutkinnonosaRepository.java
#	eperusteet-amosaa-service/src/main/java/fi/vm/sade/eperusteet/amosaa/resource/config/CrossOriginFilter.java
#	eperusteet-amosaa-service/src/test/java/fi/vm/sade/eperusteet/amosaa/service/EperusteetServiceIT.java

+++ eperusteet-amosaa-service/src/main/java/fi/vm/sade/eperusteet/amosaa/repository/koulutustoimija/custom/KoulutustoimijaRepositoryImpl.java
@@ -35,7 +35,7 @@ public class KoulutustoimijaRepositoryImpl implements KoulutustoimijaCustomRepos
         TypedQuery<Long> countQuery = getCountQuery(queryDto);
         TypedQuery<Tuple> query = getQuery(queryDto);
         if (page != null) {
-            query.setFirstResult(page.getOffset());
+            query.setFirstResult(Long.valueOf(page.getOffset()).intValue());
             query.setMaxResults(page.getPageSize());
         }
         List<Koulutustoimija> list = query.getResultList().stream()
